Вынесены конфигурации для program.cs
