+++ SurveyPlatform.API/Configuration/AddMappersConfiguration.cs
@@ -1,0 +1,12 @@
+﻿using SurveyPlatform.BLL.Mappings;
+
+namespace SurveyPlatform.API.Configuration;
+
+internal static class AddMappersConfiguration
+{
+    internal static void AddMappers(this IServiceCollection services)
+    {
+        services.AddAutoMapper(typeof(UserMapperProfile));
+        services.AddAutoMapper(typeof(UserDTOMapperProfile));
+    }
+}
